refactor(app): drop dead mirror state and demote test-only save helpers
- `GameState.services` and `GameState.tileRevision` had no production
  readers — the engine owns both concerns — so the fields, their
  factories (`createServiceSystemState` wiring), `bumpTileRevision`, and
  the `deserialize` back-fills for them are gone. `services.ts` itself
  stays: `ServiceId`/`TileServiceState` are still live.
- `serialize`/`copyState` had no production callers (the real save path
  is the binary CSAV container), so they move to
  `testSupport/legacyStateSerialize.ts` for the three test files that
  still build legacy JSON fixtures with them.
- `protocol/index.ts` re-exported the `TileKind` mapping to zero
  importers; deleted.
