--- app/src/game/protocol/index.ts
@@ -1,4 +1,0 @@
-export * from './tileKind';
-export * from './commands';
-export * from './events';
-export * from './tileBuffer';
